Add configuration for Intel oneAPI 2025.0.0 on Nautilus (#1382)
Add configuration for Intel oneAPI 2025.0.0 on Nautilus. This is with ifx, since there is no more ifort in oneapi@2025.0.0. This configuration is currently commented out, since the default is and should be 2024.2.1 with ifort for now.

This requires a bug fix in py-scipy in JCSDA/spack#483 (see there for more information and the upstream spack develop PR).