Add new template cylc-dev (and test building in GitHub actions) (#1410)
This PR adds a new template cylc-dev to build a special environment for running cylc. Because of the way cylc works, this is best accomplished by creating an environment view and then a cylc wrapper that contains the minimum settings necessary to use cylc from the view's bin directory. See cylc/cylc-flow#6532 for more information. Note that the creation of the cylc wrapper is not part of this PR.

Associated changes:
- To compile cylc-dev on the Ubuntu self-hosted runner, exclude meson from the spack external find call. Spack found a deprecated version meson@0.61 and despite it being deprecated it tried to use it - with the result that the harfbuzz build failed (needed for cylc-dev). Letting spack build meson (it picks meson@1.5.1) solved the problem.
- Remove the +excel variant from py-pandas, which was something the jedi-tools-env used, but that is no longer needed and that caused build errors (and added several more dependencies).
- Fix a bug in the GitHub actions that would allow the tests to fail silently (replace set +e with set -e).
- To avoid duplicate packages, require py-cython@3 and py-numpy@1.26
- With Intel Classic/LLVM, build qt with gcc (has no effect on pre-configured sites, where qt is configured as an external package)