Add util/nrl/batch_install.sh, add patch to build py-netcdf4 without mpi (#1464)
Add util/nrl_batch_install.sh. This includes a workaround for installing the cycl environment on Narwhal, for which we need to switch from the Cray compiler wrappers for gcc@10 to the native compilers (lines 258-264 in nrl_batch_install.sh). I will work on a better solution for this in the next weeks.

During the installation, three problems were discovered that are fixed in this PR:

1. For gcc@:10, need to use older version of cdo in configs/common/packages.yaml that uses C++-17 instead of C++-20 (same as existing modifications for Intel Classic)
2. For Atlantis, remove external (old) bison to avoi duplicate packages
3. We need to be able to build py-netcdf without MPI (~mpi) even when netcdf-c was built with +mpi to avoid errors mentioned in py-netcdf4 must be buildable with ~mpi, even if netcdf is built with ~+mpi` #1463

Also update submodule pointer for spack for changes recently merges into spack-stack-dev (py-numpy bug fix for %intel).