For release/1.9.0: cherry-pick `py-netcdf4@1.7.1: ~mpi` instead of `+mpi` from #1489 (#1491)
In PR #1489 we are changing the requirements for py-netcdf4 from py-netcdf4@1.7.1: +mpi to py-netcdf4@1.7.1: ~mpi in configs/common/packages.yaml.

This change is required to fix an error with py-netcdf4 on certain systems when built with +mpi. We used to build py-netcdf4 without mpi, but for a period this wasn't possible until we added a patch to disable the py-netcdf4 auto-detect parallel feature. That patch allows us to build py-netcdf4 ~mpi even if netcdf-c was built with +mpi.

---------

Co-authored-by: Alex Richert <alexander.richert@noaa.gov>